\name{summary.LSKAT.snp.plink}
\alias{summary.LSKAT.snp.plink}
\title{
Summarizing the LSKAT results for all SNPs
}
\description{
Summarizing the LSKAT results for all SNPs in PLINK data set
}
\usage{
summary.LSKAT.snp.plink(r.lskat)
}
\arguments{
  \item{r.lskat}{The list object with S3 class name: LSKAT.snp.plink, obtained from \code{\link{longskat_snp_plink}}.  }
}
\seealso{
\code{\link{longskat_snp_plink}} and \code{\link{summary.LSKAT.gen.plink}}
}
\keyword{ SNP-base }
\keyword{ Summary }
\keyword{ PLINK }